Move teampage_infoblock into individual folder
